pyproject.toml
[tool.poetry]
name = "pytest-testing"
version = "0.1.0"
description = ""
authors = ["Sheikh Haris Zahid <sheikh.haris.zahid@aiworks.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"

[tool.poetry.dev-dependencies]
ipython = "^8.5.0"
pytest-django = "^4.5.2"
coverage = "^6.5.0"
pytest-xdist = "^2.5.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
